ChainLadder API Reference

Please open an issue if you encounter any issues or confusion with the package.

ChainLadder.LossDevelopmentFactor

ChainLadder.age_to_age

ChainLadder.latest_diagonal

ChainLadder.linkratio

ChainLadder.square

ChainLadder.valuation_date

ChainLadder.LossDevelopmentFactor — Method

LossDevelopmentFactor(t::ClaimsTriangle;tail=nothing)

A development model using volume weighted age-to-age ratios of losses. Use log-ratios for the tail estimator unless user has specified a tail.

ChainLadder.age_to_age — Method

linkratio(t::ClaimsTriangle)

Return the age to age ratios for a given triangle.

ChainLadder.latest_diagonal — Method

latest_diagonal(t::ClaimsTriangle)

Return the last development value for each origin period.

ChainLadder.linkratio — Method

linkratio(t::ClaimsTriangle)

Return the link ratios for a given triangle.

ChainLadder.square — Method

Fill in, or "square" the triangle with a fitted model.

ChainLadder.valuation_date — Method

valuation_date(t::ClaimsTriangle)

Return the last development index.